minimally tweak UI elements and make Microsoft Visual Studio :: XAML Designer *NOT* barf a hairball (Fatal System Exception) on many XAML dialogs/panels in Qiqqa.
The key to the latter is tweaked the dialog setup/init code such that XDevProc (which is part of Visual Studio) can execute that part of the Qiqqa code and thus render the panel -- or at least not b0rk and fail on it when you open the XAML file in the Designer.

---

WARNING NOTE:

TODO: The bit at the end is a copy/dump off SO where the same problem as https://github.com/jimmejardine/qiqqa-open-source/issues/40

Must be tweaked and then retried with Qiqqa as it didn't fly for me yet. :-(

The blunter hack of tweaked the Qiqqa code and sneakily support XDesProc code flow too
by moving the Theme loader into Utilities.GUI is nice, but somehow still makes Designer
blue-wave the Brushes and NOT render them.

---

```
<UserControl.Resources>
    <ResourceDictionary>
        <ResourceDictionary.MergedDictionaries>
			<ResourceDictionary Source="pack://application:,,,/Styles;component/Resources.xaml" />
        </ResourceDictionary.MergedDictionaries>
    </ResourceDictionary>
</UserControl.Resources>
```