[FEATURE] work done for https://github.com/jimmejardine/qiqqa-open-source/issues/21 : the BibTeX entry definitions are now sitting in an external JSON file so every savvy user can edit/add their own set of fields there now.
BTW: Nunit testing was a great idea but I'm getting savagely abused around my Glutus Maximus with a Cactus by way of the woes of https://stackoverflow.com/questions/15614192/run-a-specified-nunit-test-as-a-32-bit-process-in-a-64-bit-environment . I effin' *hate* .NET right now - and particularly old cruft that needs upgrading so I can move out of that fugly 32bit shyte sewer.      \< angry like mad! />