refactor Daemon class and associated background tasks: make it possible for registered tasks to be *removed* properly: this is useful as the load/init phase of Qiqqa MAY otherwise exhibit multiple folderWatcher registrations for the same library, once for each time it is updated during the initial scan-for-suitable-libraries phase of the application.
Note: The code is simplified by never re-using a slot in the daemon registration list: this does no harm to performance nor memory consumption as a *large* test environment (my own rig) clocks in at only 87 slots used during an entire run, while the test rig comes with 20+ actual libraries and various 'paths' through which the libraries will be discovered during the scan-and-load init/load phase of the application. Hence we expect even extreme and heavy users of Qiqqa will not surpass about 100 slots in the registration array/list. Thus we have designed the code to use the simplest approach for the registration array/list: grow-only, NULLing the discarded slots.