ran into serious trouble today: the interprocess I/O over stdout for mutool page rendering failed all of a sudden. Spent quite some time checking out what went wrong and how to fix it.
Trouble is I don't know why the original code suddenly FAILED. The only change I made that COULD impact in any way was the change from building the mupdf tools in 64 bit mode instead of 32-bit mode when building Debug/Release mixed mode builds. This, however, SHOULD NOT have changed the interproceess behaviour that way (or in ANY way)!

Anyway, added a System Test to verify proper behaviour of that bit of code as it's been brittle in the past and would benefit greatly from some system tests, rather than having to test this in a complex multi-threaded environment like Qiqqa where these calls happen in parallel and are therefor a *devil* to debug!

Used the code from Cysharp/ProcessX (https://github.com/Cysharp/ProcessX) as a base for the re-engineering of the exeec/spawn functionality.