Further work done on https://github.com/jimmejardine/qiqqa-open-source/issues/328 : Portable Application Mode.
In that mode, the registry is only used for a few things, such as registering the qiqqa:// URI and qiaa_backup file extension for qiqqa backup archives. Otherwise, all settings (such as Qiqqa Base Directory) are obtained from the "Qiqqa.Portable.Settings.json5" JSON5-formatted configuration file, that must be present in the same directory as the qiqqa.exe executable for it to be recognized as a Portable Application.

The technique used is to make a Portable Application a kind of "developer override" mode: the same files are involved and the way configuration bits and pieces are specified/overridden is the same.

Re "developer override mode":

Now we support the "Qiqqa.Developer.Settings.json5" developer config file in the application directory too! (Next to that "Qiqqa.Portable.Settings.json5" file)

When we have determined the Base Directory in Qiqqa, we also load the *other* "Qiqqa.Developer.Settings.json5" file that may be available there: hence we now support a chain of 2..3 developer override config files, all formatted the same way, and loaded in this order

- "Qiqqa.Portable.Settings.json5"   in Qiqqa.exe's directory
- "Qiqqa.Developer.Settings.json5"  in Qiqqa.exe's directory
- (determine Base Directory by checking the files above and optionally, the Windows Registry)
- "Qiqqa.Developer.Settings.json5"  in the Base Directory

All these files are optional; the EXISTANCE of the "Qiqqa.Portable.Settings.json5" file determines whether Qiqqa acts like a Portable Application or as a regular install - the file may even be empty then!

Extra: the developer config files can now also override the registry entries used by Qiqqa for various configuration settings:

{
  "DebugConsole": false,
  "AllowMultipleQiqqaInstances": true,
  "BaseDataDirectory": "Z:\\lib\\tooling\\qiqqa\\Qiqqa\\bin\\Debug\\My.Qiqqa.Libraries",
  "LoadKnownWebLibraries": true,
  "AddLegacyWebLibrariesThatCanBeFoundOnDisk": true,
  "SaveKnownWebLibraries": true,
  "DoInterestingAnalysis_GoogleScholar": false,
  "FolderWatcher": true,
  "TextExtraction": true,
  "SuggestingMetadata": true,
  "BuildSearchIndex": true,
  "RenderPDFPagesForSidePanels": true,
  "RenderPDFPagesForReading": true,
  "RenderPDFPagesForOCR": true,
  "FirstInstallNotification": "83"
}