remove code to set particular images' rendering to high quality mode: now all images are rendered in efault hardware mode (LowQuality) - it's just the upscaled ones that look crappy.
(Note: this now clearly shows the page renders to be low grade; meanwhile the render cost in MuPDF has a high baseline independent of pixel real estate size, so the added cost of rendering larger images there is noticable, but not detrimental. And if it turns out to be, we'ld better addreess the issue there for I want crisp, readable pages. ;-)