tweak the abstract fetching code for the popup-on-hover UI bit in the library list: don't use magic strings to recognize state.
TODO: here, as elsewhere in the code, we now NEED a better approach to properly cope with the rather more "progressive refinement" approach I want to arrive at in the UI.
This includes picking up cached images of lower rez, upscaling them until a higher rez becomes available.

Introduced the first minimal bit of work for this: the ProgressiveContent template class.

(And now I had to read up on WPF and how to manipulate that bugger to do my bidding. Turns out it should've been easy were it not for the nuked not-MVVM Qiqqa GUI. !@#$%^&*#)

More work to do.

Also got rid of the legacy WPF straight-copy data converter.