slowly untangling the mess that is the UI code around the PDF viewer.
Also added code to override the PDF page rendering process and plug in a few "still pending" images instead. This helps us test the UI and will also assist later when we get to the point where we introduce the "progressive updates"...

Currently it's still the old mess and oddly enough the "pending" images don't show up yet either, so it's already good testing material.  :'-(