re-organizing doc-src/ -- following first experiences with Obsidian (which is very useful indeed otherwise), it turns out to be better (better Obsidian performance, no black screen bug trigger, etc.etc.) to keep Obsidian Vaults as a kind of "end node" in your directory tree. Hence anything that was previously not a fully integral part of the dev notes but still sat inside doc-src has been re-organized into separate directory "end nodes" themselves.
While we work on deGaulle (intermittantly; our site generator) that tool will have to be adjusted accordingly as these new insights are important: Obsidian does node graph visualization and traversal very well, which I've been looking for for a long time and never found a good (enough) solution for. deGaulle is a from-scratch effort to make it happen at a website viewing & authoring level, but Obsidian is there and doing what needs to happen (site-authoring-wise) so deGaulle will have to adjust to suit Obsidian instead of vice-versa.

Of course, encountering Obsidian has led to discovering several other (open source) tools that purportedly offer the same, but those have to be investigated another day; initial checks on a few didn't produce a satisfactory match there, so Obsidian it is.