refactor PDF bookmarks menu builder to properly use UI and background threads.
TODO: menu *could* be optimized from a UX perspective by showing a preliminary menu while the full one is contructed from extracted PDF info in the background.
Management Decision: we'll leave that sort of fanciness to the next UI (webview based) as doing this in WPF is too much hassle.