QiqqaOCR: (see also previous commit) *only disabling the FONT-CHANGE metric* when the following condition is NOT met: both sides are letter/digit/full-ASCII. This would thus negate the font-change condition and merge TeX-produced SmallCaps words, for example: both sides are 'letter' or 'digit' then!
Postscript: this completely fixes the test file's Reference section. Of course, the inline math gets buggered in a few places, merging with the next follow-up word in the regular text (due to switching from a overlapping cursive font to regular and lots of kerning) but I deem this quite acceptable right now as almost always you can expect low-importance words like 'assumes' and stop words like 'is', 'and', etc. to follow an in-line math expression, so that bit of corruption is tolerable in the output. Meanwhile all SmallCaps and special keywords are now bundled properly into words, thus improving our chancee of success of finding those once indexed in our FTS engine.