Ok, so currently the PDF "Properties" panel is also crashing. Given today's hunt & finding that, on this dev box at least, Unicode characters CAN break WPF internals fatally (see other commits of today), it sounds & smells a lot like https://github.com/AvaloniaUI/Avalonia/issues/4755 but I DO NOT USE AVALONIA. Apparently WPF itself is similarly crappy in this regard.
I haven't touched Qiqqa C# code in a few months, but the only real change to the dev machine is a MSVC2022 install - besides a complete hardware change before that, but the "fun" bit of that story is that Qiqqa DID RUN CORRECTLY before I installed MSVC2022 on the new hardware. Given that DevStudio looks *slightly* different from before (also MSVC2019) I wouldn't be surprised if there's a system level "font update" between "was working here in MSVC2019" and "b0rks horridly for me now in MSVC-whhatever!!1!!" states of the machinery.

A fuckin' nightmare to debug, that's for sure, taking ages to drill down to root cause. I can't wait to get rid of this XAML/WPF shite.